Moved the run_foo_test functions to a separate file
